hotfix: reorganized resend service exposing to public causes crashing browser
